# specialized FixedIndexed implementations for java value types