… processing/src/main/java/org/apache/druid/segment/nested/CompressedNestedDataComplexColumn.java
… processing/src/main/java/org/apache/druid/segment/nested/NestedDataColumnSerializer.java
… processing/src/main/java/org/apache/druid/segment/nested/NestedDataColumnSupplier.java
… processing/src/main/java/org/apache/druid/segment/nested/NestedFieldLiteralColumnIndexSupplier.java
… processing/src/main/java/org/apache/druid/segment/nested/NestedFieldLiteralDictionaryEncodedColumn.java
… processing/src/test/java/org/apache/druid/segment/data/FixedIndexedTest.java
… processing/src/test/java/org/apache/druid/segment/nested/NestedFieldLiteralColumnIndexSupplierTest.java